+++ netclop/__init__.py
@@ -1,4 +1,5 @@
-from .networkensemble import NetworkEnsemble
+"""Package initialization."""
 from .geo import GeoNet, GeoPlot
+from .networkensemble import NetworkEnsemble
 from .sigclu import SigClu
 from .upsetplot import UpSetPlot
