+++ netclop/constants.py
@@ -1,7 +1,4 @@
 """Defines constants."""
-type Node = str
-type Partition = list[set[Node]]
-
 COLORS = ["#636EFA", "#EF553B", "#00CC96", "#FFA15A", "#AB63FA",
           "#19D3F3", "#FF6692", "#B6E880", "#FF97FF", "#FECB52"]
