+++ netclop/typing.py
@@ -1,0 +1,4 @@
+"""Defines types."""
+type Cell = int
+type Node = str
+type Partition = list[set[Node]]
